+++ src/main/java/edu/ie3/datamodel/io/factory/input/Transformer2WInputFactory.java
@@ -37,6 +37,13 @@ public class Transformer2WInputFactory
       NodeInput nodeB,
       OperatorInput operator,
       OperationTime operationTime) {
+
+    if (nodeB.getVoltLvl().getNominalVoltage().compareTo(nodeA.getVoltLvl().getNominalVoltage())
+        > 0) {
+      throw new IllegalArgumentException(
+          "nodeA must be on the high voltage side of the transformer");
+    }
+
     final int parallelDevices = data.getInt(PARALLEL_DEVICES);
     final Transformer2WTypeInput type = data.getType();
     final int tapPos = data.getInt(TAP_POS);