+++ src/main/java/edu/ie3/datamodel/models/input/connector/Transformer3WInput.java
@@ -304,6 +304,15 @@ public class Transformer3WInput extends TransformerInput implements HasType {
 
     @Override
     public Transformer3WInput build() {
+
+      NodeInput nodeA = super.getNodeA();
+      NodeInput nodeB = super.getNodeB();
+
+      if (nodeB.getVoltLvl().getNominalVoltage().compareTo(nodeA.getVoltLvl().getNominalVoltage())
+          > 0) {
+        throw new IllegalArgumentException(
+            "nodeA must be on the high voltage side of the transformer");
+      }
       return new Transformer3WInput(
           getUuid(),
           getId(),